requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.5
Django==4.0.6
django-allauth==0.51.0
django-cache-url==3.4.2
django-countries==7.3.2
django-crispy-forms==1.14.0
django-phonenumber-field==6.3.0
environs==9.5.0
idna==3.3
marshmallow==3.17.0
oauthlib==3.2.0
packaging==21.3
phonenumbers==8.12.52
phonenumberslite==8.12.52
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
